--- src/Finbuckle.MultiTenant.Contrib.EFCoreStore/EFCacheStoreConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace Finbuckle.MultiTenant.Contrib.EFCoreStore
-{
-    public partial class EFCacheStoreConfiguration
-    {
-        public int CacheMinutes { get; set; }
-    }
-}
