+++ src/Finbuckle.MultiTenant.Contrib.IdentityServer/README.md
@@ -1,0 +1,2 @@
+# Finbuckle.MultiTenant.Contrib
+Contribution functionality for [Finbuckle.MultiTenant](https://www.finbuckle.com/MultiTenant).
