--- src/Finbuckle.MultiTenant.Contrib/TenantClaimConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace Finbuckle.MultiTenant.Contrib.Abstractions
-{
-    public class TenantClaimConfiguration
-    {
-        public string Name { get; set; }
-    }
-}
